// src/utils/inicio/types.ts
export interface ILocal {
  id: number;
  definicaoLocalForm: string;
}
export interface IGrau {
  id: number;
  definicaoGrau: string;
}

export interface IArea {
  id: number;
  definicaoArea: string;
}

export interface ITipo {
  id: number;
  definicaoTipo: string;
}

export type Users = {
  id: number;
  name: string;
};